Fix no-signoff feature for user-owned repositories
When the application is installed on a repository that is owned
by a regular user (as opposed to an organizaton) we now check if
the commit author is the same as the repository owner, and allow
skipping the sign-off only then.

This also fixed a bug that broke DCO on repositories owned by
users. The bug was introduced by #52.

Signed-off-by: Lóránt Pintér <lorant.pinter@gmail.com>