Add link to check out PR locally
When a user is making changes via the GitHub web UI (for example, editing a
README.md in their browser using the edit button) they may not have a local
copy of the branch. This edits the failure message to include a link to
instructions on how to clone a PR locally.

Signed-off-by: Brian Warner <brian@bdwarner.com>