Add full support for remediation commits
This commit adds support for individual and third-party remediation commits. These are disabled
unless explicitly enabled by the end user to preserve the default, more conservative functionality.

* What is a remediation commit?

In brief, a remediation commit is a way to achieve DCO signoffs without amending the original
commit. Remediation commits were proposed upon feedback from open source communities that PRs were
failing or being abandoned because not everybody checks their code out to a local git environment.
In particular, the GitHub UI does not currently support amending a commit log, so PRs which failed
the Probot DCO check were being abandoned by committers. Remediation commits are proposed as an
improvement because they can be executed entirely within the GitHub web UI, while still being a part
of a project's immutable and portable git history.

* Why would I use this when I can amend commits?

The process of amending commits works well for single contributors working alone in a branch, but it
has challenges when there are multiple contributors or when rebasing the branch would break others'
workflows. For example, `git rebase HEAD~10 --signoff` adds the author's signoff to all 10 prior
commits, whether they authored them or not.

The advantage of this approach is that it allows contributors to pinpoint the exact commits for
which they need to add a signoff, even if they've already been merged.

* How does a remediation commit work?

There are two types of remediation commits: individual and third-party.

For an individual remediation, the original author pushes a second commit with specific text in the
commit log indicating they are adding their Signed-off-by line retroactively. Only the original
author can provide an individual remedation.

For a third-party remediation, someone else pushes a second commit with specific text in the commit
log indicating they are adding their Signed-off-by line in place of the author. This is intended for
situations where the original author is not available and where another can make the DCO commitment
in their place (e.g., an employer).

* Why is this not turned on by default?

The default Probot DCO behavior mirrors the workflow of the Linux kernel, where an author who
forgets a DCO signoff must amend their commits and resubmit them. This is a well-known process and
a broadly adopted requirement, and the intent is not to change how Probot works for projects that
have adopted it under these expectations.

* How do I enable this?

Projects can enable remediation commits on a repo-by-repo basis by adding a `dco.yml` config file in
their `/.github` directory with the following lines:

  allowRemediationCommits:
    individual: true
    thirdParty: true

Note that individual remediations must be enabled for thirdParty remediations to work.

* When should I enable this?

You should have a conversation with your project leadership to determine "if remediation commits are
sufficient to be considered a DCO Sign-off." There may be different answers for different projects.
Contact your legal counsel if you aren't sure what this means.

* How does a contributor know what remediation text to use?

Much as the DCO signoff text is standardized ("Signed-off-by: Name <email>"), the individual and
third-party remediations are also intended to be standardized. This will allow other tools to search
for the text and come to the same conclusion whether a commit has been correctly signed off. When
the check runs, it will show the exact text the author or a third-party will need to copy into the
commit message to constitute a valid remediation commit.

In general, the standard signoff texts are:
Default: Signed-off-by: Brian Warner <brian@bdwarner.com>
Individual: I, Brian Warner <brian@bdwarner.com>, hereby add my Signed-off-by to this commit: <SHA>
Third-party: On behalf of Not Brian <notbrian@notbdwarner.com>, I, Brian Warner <brian@bdwarner.com>, hereby add my Signed-off-by to this commit: <SHA>

---

This commit includes the following:

* Support for individual remediation commits
* Support for third-party remediation commits
* Tests that verify the new behavior
* Tests that verify the default behavior is unchanged when remediation commits are enabled
* An additional warning about how rebasing can break workflows if remediation is available

Signed-off-by: Brian Warner <brian@bdwarner.com>